+++ app/Http/Controllers/PartidoController.php
@@ -21,24 +21,58 @@ class PartidoController extends Controller
      */
     public function index()
     {
+      if (\Auth::user()->tipo == "admin")
+      {
         $jornadas = Jornada::all();
         $partidos = Partido::all();
         $equipos = Equipo::all();
 
         return view('partidos.partidosIndex', compact('partidos', 'jornadas', 'equipos'));
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA USUARIO
+      }
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @param  \App\Jornada  $jor
+     * @param  \App\Jornada  $jornada
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create(Jornada $jornada)
     {
+      if(\Auth::user()->tipo == "admin")
+      {
         $jornadas = Jornada::all();
         $equipos = Equipo::all();
-        return view('partidos.partidoForm', compact('jornadas', 'equipos'));
+        return view('partidos.partidoForm', compact('jornadas', 'equipos',  'jornada'));
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA index
+      }
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @param  \App\Jornada  $jornada
+     * @return \Illuminate\Http\Response
+     */
+    public function createParJorX(Jornada $jornada)
+    {
+      if(\Auth::user()->tipo == "admin")
+      {
+        $jornadas = Jornada::all();
+        $equipos = Equipo::all();
+        return view('partidos.partidoForm', compact('jornadas', 'equipos', 'jornada'));
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA index
+      }
     }
 
     /**
@@ -49,6 +83,8 @@ class PartidoController extends Controller
      */
     public function store(Request $request)
     {
+      if(\Auth::user()->tipo == "admin")
+      {
         $request->validate([
           'jornada_id' => 'required|integer',
           'equipo_local' => 'required|integer',//|different:equipo_visitante'
@@ -85,6 +121,11 @@ class PartidoController extends Controller
         */
 
         return redirect()->route('partidos.index');
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA INEDEx
+      }
 
     }
 
@@ -96,23 +137,37 @@ class PartidoController extends Controller
      */
     public function show(Partido $partido)
     {
+      if (\Auth::user()->tipo == "admin")
+      {
         $jornadas = Jornada::all();
         $equipos = Equipo::all();
         return view('partidos.partidoShow', compact('partido', 'jornadas', 'equipos'));
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA USUARIO
+      }
     }
 
     /**
      * Muestra los partidos de X jornada
      *
-     * @param  \App\Partido  $partido
+     * @param  \App\Jornada  $jornada
      * @return \Illuminate\Http\Response
      */
     public function showJorn(Jornada $jornada)
     {
+      if (\Auth::user()->tipo == "admin")
+      {
         $jornadas = Jornada::all();
         $partidos = Partido::all();
         $equipos = Equipo::all();
         return view('partidos.partidosJornada', compact('partidos', 'jornadas', 'equipos', 'jornada'));
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA USUARIO
+      }
     }
 
     /**
@@ -123,9 +178,16 @@ class PartidoController extends Controller
      */
     public function edit(Partido $partido)
     {
+      if (\Auth::user()->tipo == "admin")
+      {
         $jornadas = Jornada::all();
         $equipos = Equipo::all();
         return view('partidos.partidoForm', compact('partido', 'equipos', 'jornadas'));
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA index
+      }
     }
 
     /**
@@ -136,10 +198,17 @@ class PartidoController extends Controller
      */
     public function editRes(Jornada $jornada)
     {
+      if (\Auth::user()->tipo == "admin")
+      {
         $jornadas = Jornada::all();
         $partidos = Partido::all();
         $equipos = Equipo::all();
         return view('partidos.partidoValoresForm', compact('partidos', 'jornadas', 'equipos', 'jornada'));
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA index
+      }
     }
 
     /**
@@ -151,24 +220,31 @@ class PartidoController extends Controller
      */
     public function update(Request $request, Partido $partido)
     {
-      $request->validate([
-        'jornada_id' => 'required|integer',
-        'equipo_local' => 'required|integer',//different:equipo_visitante
-        'equipo_visitante' => 'required|integer|different:equipo_local',
-        'resL' => 'nullable|integer',
-        'resV' => 'nullable|integer',
-      ]);
+      if (\Auth::user()->tipo == "admin")
+      {
+        $request->validate([
+          'jornada_id' => 'required|integer',
+          'equipo_local' => 'required|integer',//different:equipo_visitante
+          'equipo_visitante' => 'required|integer|different:equipo_local',
+          'resL' => 'nullable|integer',
+          'resV' => 'nullable|integer',
+        ]);
 
-      $partido->jornada_id = $request->jornada_id;
-      $partido->equipo_local = $request->equipo_local;
-      $partido->equipo_visitante = $request->equipo_visitante;
-      //$partido->resL = 0;
-      //$partido->resV = 0;
-      $partido->resL = $request->resL;
-      $partido->resV = $request->resV;
-      $partido->save();
+        $partido->jornada_id = $request->jornada_id;
+        $partido->equipo_local = $request->equipo_local;
+        $partido->equipo_visitante = $request->equipo_visitante;
+        //$partido->resL = 0;
+        //$partido->resV = 0;
+        $partido->resL = $request->resL;
+        $partido->resV = $request->resV;
+        $partido->save();
 
-      return redirect()->route('partidos.show', $partido->id);
+        return redirect()->route('partidos.show', $partido->id);
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA index
+      }
     }
 
     /**
@@ -181,16 +257,23 @@ class PartidoController extends Controller
      */
     public function updateRes(Request $request, Partido $partido, Jornada $jornada)
     {
-      $request->validate([
-        'resL' => 'nullable|integer',
-        'resV' => 'nullable|integer',
-      ]);
+      if (\Auth::user()->tipo == "admin")
+      {
+        $request->validate([
+          'resL' => 'nullable|integer',
+          'resV' => 'nullable|integer',
+        ]);
 
-      $partido->resL = $request->resL;
-      $partido->resV = $request->resV;
-      $partido->save();
+        $partido->resL = $request->resL;
+        $partido->resV = $request->resV;
+        $partido->save();
 
-      return redirect()->route('partidos.editRes', $partido->jornada_id);
+        return redirect()->route('partidos.editRes', $partido->jornada_id);
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA index
+      }
     }
 
     /**
@@ -201,7 +284,14 @@ class PartidoController extends Controller
      */
     public function destroy(Partido $partido)
     {
+      if (\Auth::user()->tipo == "admin")
+      {
         $partido->delete();
         return redirect()->route('partidos.index');
+      }
+      elseif(\Auth::user()->tipo == "comun")
+      {
+        //VISTA index
+      }
     }
 }
